+++ phytreon/__init__.py
@@ -39,7 +39,7 @@ from .infer import (
 from .comparative import ace_parsimony, ace_ml, ace_continuous, stochastic_map
 from .treeops import (
     rotate, flip, swap_children, ladderize, sort_by, collapse_low_support,
-    scale_clade, cut_tree, midpoint_root, group_clade, group_otu,
+    scale_clade, cut_tree, midpoint_root, outgroup_root, group_clade, group_otu,
     robinson_foulds, prune_to_taxa, untangle, crossing_number,
     collapse_clade,
 )
@@ -62,6 +62,6 @@ __all__ = [
     "expression_distance_matrix", "expression_dendrogram",
     "ace_parsimony", "ace_ml", "ace_continuous", "stochastic_map",
     "rotate", "flip", "swap_children", "ladderize", "sort_by", "collapse_low_support",
-    "scale_clade", "cut_tree", "midpoint_root", "group_clade", "group_otu",
+    "scale_clade", "cut_tree", "midpoint_root", "outgroup_root", "group_clade", "group_otu",
     "prune_to_taxa", "untangle", "crossing_number", "collapse_clade",
 ]